backend/requirements.txt
fastapi 
uvicorn[standard]
sqlalchemy
python-dotenv
psycopg2 #for postgres
asyncpg #for postgres